Fixed Ubuntu/Darwin difference for size_t/uint64.
Updated template<class> struct extrema; to use requires clauses for
specialization. This eliminates dependencies on particular types and
unifies the treatment of size_t/uint64 on Ubuntu/Darwin.

Updated hash and hasher to consistently use uint64 eliminated mention
of size_t.

+++ include/core/hash/hash.h
@@ -8,8 +8,8 @@ namespace core {
 
 // hash - return hash of arguments
 template<class... Args>
-size_t hash(Args&&... args) {
-    size_t v{0};
+uint64 hash(Args&&... args) {
+    uint64 v{0};
     (detail::combine(v, detail::hasher<std::decay_t<Args>>()(std::forward<Args>(args))),...);
     return v;
 }
